# docs/openapi.yaml
openapi: 3.1.0
info:
  title: sms-hub API
  version: 2.0.0
  summary: 瘦终端 + Hub 短信网关的完整 HTTP 契约(v2)
  description: |
    三个通信平面:
      1. **设备 → Hub**:`POST /hook/{token}`(设备主动上报,token 即全局 DEVICE_TOKEN)。
      2. **浏览器 → Hub**:`/api/*`(除 health/login/metrics 外需 `Authorization: Bearer <token>`)。
      3. **Hub → 设备**:`/{token}/pull|send|delete|at|update`(由瘦终端固件提供,路径级 `servers` 覆盖指向设备;
         token 不匹配返回 404 空响应;请求体必须 `ensure_ascii=False`)。

    卡片为中心:`sim_id`(IMSI 派生)是短信/发件/通知/收件筛选的业务主键;`device_mac` 是承载载体。
    当前卡片类 API 的 `sim_id` 缺省时,恰 1 张启用卡自动推断,0 或 >1 张时返回 400。

    与本文对齐的人类可读文档:`docs/guide.md`(设计与接口权威:架构/可靠性/接口契约/数据模型)。
    本文件是机器可读、可校验的契约来源;前端类型应据此生成以杜绝手写漂移。

  contact: {}
  license:
    name: MIT
servers:
  - url: http://{hubHost}
    description: Hub —— 浏览器与设备上报入口
    variables:
      hubHost:
        description: Hub 监听地址:端口(默认 :8025)
        default: 127.0.0.1:8025

# 默认所有 /api/* 需 Bearer;公开端点用 security: [] 覆盖
security:
  - bearerAuth: []
...
paths:
...
  /api/status:
    get:
      tags: [Status & Control]
      summary: 当前卡片 + 承载设备合成状态(读心跳缓存,不打扰设备)
      parameters:
        - $ref: '#/components/parameters/SimIdQuery'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StatusResponse'}
        '400':
          $ref: '#/components/responses/Error'
  /api/poll:
    post:
      tags: [Status & Control]
      summary: 强制拉取一批入库
      description: 立即对该卡片承载设备拉一轮(PULL_BATCH=20),返回新入库条数。
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SimBody'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PollResponse'}
        '409':
          $ref: '#/components/responses/Error'
        '502':
          $ref: '#/components/responses/Error'
  /api/status/refresh:
    post:
      tags: [Status & Control]
      summary: 刷新 modem 身份(pull?include_status=1, limit=1)
      description: 与 /api/poll 不同:只刷新 modem 身份块,不批量入库,返回 age_s=0。
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SimBody'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RefreshResponse'}
        '409':
          $ref: '#/components/responses/Error'
        '502':
          $ref: '#/components/responses/Error'
  /api/buffer/clear:
    post:
      tags: [Status & Control]
      summary: 排空设备缓冲里已同步到 Hub 的消息
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SimBody'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ClearBufferResponse'}
        '409':
          $ref: '#/components/responses/Error'
        '502':
          $ref: '#/components/responses/Error'
  /api/at:
    post:
      tags: [Status & Control]
      summary: AT 透传(不排队,设备忙即 409)
      description: |
        超时 timeout_ms 限制在 100..15000。**不排队**:设备正忙或已有短信任务等待时直接 409。
        响应直接透传设备 `{ok, response}`(无独立 error 字段);失败原因(如 `TIMEOUT`、`缺少 cmd 字段`)放 response,ok=false。
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AtBody'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AtResponse'}
        '409':
          $ref: '#/components/responses/Error'
        '502':
          $ref: '#/components/responses/Error'

  # ════════════ Config & Notify ════════════
...
  /hook/{token}:
    post:
      tags: [Device-Webhook]
      summary: 设备上报入口(门铃 + 地址上报 + 状态快照)
      description: |
        token 只认证、不区分设备,body 必须带 `mac`。token 不匹配返回 **404 空响应**。
        不在请求内同步 pull —— 只做 token 校验、MAC/IMSI 规范化、状态快照保存、异步拉取排队。
        ip:port 须经 SSRF 校验(仅 RFC1918/ULA)才写入 base_url。
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema: {type: string}
          description: 全局 DEVICE_TOKEN
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WebhookBody'}
            examples:
              heartbeat:
                summary: heartbeat/boot 携带完整状态(含 modem.imsi)
                value:
                  event: heartbeat
                  mac: aabbccddeeff
                  seq_id: 12
                  ip: 192.168.1.88
                  port: 80
                  fw: 2.1.0
                  heartbeat_interval_s: 60
                  modem:
                    ready: true
                    operator: CMCC
                    csq_dbm: -78
                    imsi: '460001234567731'
                    imsi_tail: '7731'
                    msisdn: '13800138000'
                    iccid: 898600A00000F0217731
                    iccid_tail: '7731'
                  buffer: {oldest_id: 88, latest_id: 123, count: 4, capacity: 50, dropped_total: 0}
              sms:
                summary: sms/hello 只带轻量字段
                value:
                  event: sms
                  mac: aabbccddeeff
                  ip: 192.168.1.88
                  port: 80
                  buffer: {latest_id: 124}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: {type: boolean}
                  pull_scheduled: {type: boolean, description: 本轮是否排了异步拉取}
                required: [ok, pull_scheduled]
        '400':
          description: MAC 缺失/非法,或 ip:port 未过 SSRF 校验
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        '404':
          description: token 不匹配(空响应)

  # ════════════ Hub → Device(设备固件提供,路径级 servers 覆盖)════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 会话令牌(POST /api/login 获取),TTL 30 天。
  parameters:
    SimIdQuery:
      name: sim_id
      in: query
      required: false
      schema: {type: string}
      description: '具体 sim_id / `all`(全部卡) / `online`(仅在线卡);缺省时单卡自动推断,多卡须显式(否则 400)'
...
  responses:
    Error:
      description: 错误
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
    Error:
      type: object
      properties:
        detail: {type: string}
      required: [detail]
...
    WebhookBody:
      type: object
      required: [event, mac]
      properties:
        event: {type: string, enum: [boot, hello, sms, heartbeat]}
        mac: {type: string, description: 小写无分隔 12 位 hex(也接受冒号/横线形式)}
        seq_id: {type: integer, description: 同物理设备内单调递增序号}
        device_ts_ms: {type: integer}
        ip: {type: string, description: 设备自报 IP,须经 SSRF 校验}
        port: {type: integer, minimum: 1, maximum: 65535}
        fw: {type: string}
        heartbeat_interval_s: {type: integer}
        uptime_s: {type: integer}
        modem: {$ref: '#/components/schemas/ModemSnapshot'}
        buffer: {$ref: '#/components/schemas/BufferSummary'}
        counters: {$ref: '#/components/schemas/Counters'}
        last_error: {type: string}
    ModemSnapshot:
      type: object
      properties:
        ready: {type: boolean}
        model: {type: string}
        cereg: {type: integer}
        csq_dbm: {type: integer}
        operator: {type: string}
        sim: {type: boolean}
        imsi: {type: string, description: 完整 IMSI(仅受控通道瞬时用,Hub 不持久化/回传)}
        imsi_tail: {type: string}
        msisdn: {type: string}
        iccid: {type: string}
        iccid_tail: {type: string}
        apn: {type: string}
        flight_mode: {type: integer}
        pdp_active: {type: boolean}
    BufferSummary:
      type: object
      properties:
        oldest_id: {type: integer}
        latest_id: {type: integer}
        count: {type: integer}
        capacity: {type: integer}
        dropped_total: {type: integer}
        last_pulled_id: {type: integer, description: 仅设备 /status 快照含}
    Counters:
      type: object
      properties:
        rx_total: {type: integer}
        tx_total: {type: integer}
        webhook_fail_total: {type: integer}

    # ── 设备/卡视图 ──
...
    SimBody:
      type: object
      properties:
        sim_id: {type: string, nullable: true}
...
    AtBody:
      type: object
      required: [cmd]
      properties:
        sim_id: {type: string, nullable: true}
        cmd: {type: string}
        timeout_ms: {type: integer, minimum: 100, maximum: 15000, default: 3000}

    # ── 消息 ──
...
    StatusHub:
      type: object
      required: [sim_id, sim_name, device_mac, stored_total, cursor, last_poll_ago_s, last_hook_ago_s, poll_interval_s, device_busy]
      properties:
        sim_id: {type: string}
        sim_name: {type: string}
        device_mac: {type: string}
        stored_total: {type: integer}
        cursor: {type: integer}
        last_poll_ago_s: {type: integer}
        last_hook_ago_s: {type: integer}
        poll_interval_s: {type: integer}
        device_busy: {type: string}
    StatusResponse:
      type: object
      required: [device, device_reachable, overall_online, heartbeat_online, data_plane_online, device_status_age_s, hub]
      properties:
        device: {$ref: '#/components/schemas/DeviceStatus'}
        device_reachable: {type: boolean, description: overall_online 的兼容别名}
        overall_online: {type: boolean}
        heartbeat_online: {type: boolean}
        data_plane_online: {type: boolean}
        device_status_age_s: {type: integer}
        hub: {$ref: '#/components/schemas/StatusHub'}
    PollResponse:
      type: object
      required: [ok, sim_id, device_mac, inserted]
      properties:
        ok: {type: boolean}
        sim_id: {type: string}
        device_mac: {type: string}
        inserted: {type: integer}
    RefreshResponse:
      type: object
      required: [ok, sim_id, device_mac, age_s]
      properties:
        ok: {type: boolean}
        sim_id: {type: string}
        device_mac: {type: string}
        age_s: {type: integer}
    ClearBufferResponse:
      type: object
      required: [ok, sim_id, device_mac, deleted]
      properties:
        ok: {type: boolean}
        sim_id: {type: string}
        device_mac: {type: string}
        deleted: {type: integer}
    AtResponse:
      type: object
      description: 透传设备响应;失败原因放 response,无 error 字段
      required: [ok, response]
      properties:
        ok: {type: boolean}
        response: {type: string}

    # ── 配置/通知 ──
...
    DeviceStatus:
      type: object
      description: 设备完整快照(heartbeat/boot body 或 /status 响应)
      properties:
        fw: {type: string}
        uptime_s: {type: integer}
        free_heap: {type: integer}
        min_free_heap: {type: integer}
        wifi_rssi: {type: integer}
        wifi_ssid: {type: string}
        reset_reason: {type: string}
        last_error: {type: string}
        heartbeat_interval_s: {type: integer}
        modem: {$ref: '#/components/schemas/ModemSnapshot'}
        buffer: {$ref: '#/components/schemas/BufferSummary'}
        counters: {$ref: '#/components/schemas/Counters'}